+++ CHANGELOG.md
@@ -4,6 +4,12 @@ Change Log
 2017-06-24
 ----------
 
+ * changed property name from email to value in EmailAddress (more consistent)
+ * added PhoneNumber VO
+ 
+2017-06-24
+----------
+
  * added hasKey, hasValue and values to AbstractEnumeration
 
 2017-06-23